+++ src/main.rs
@@ -1,13 +1,7 @@
-#[macro_use]
-extern crate nom;
-
-// Import helper macros before `parsers`
-#[macro_use]
-mod helper_parsers;
-
 mod error;
 mod file_utils;
 mod parsers;
+mod parsing_types;
 mod transpile;
 
 enum ReturnCode {